+++ axum-extra/src/response/mod.rs
@@ -12,6 +12,13 @@ pub mod multiple;
 #[cfg(feature = "error-response")]
 mod error_response;
 
+#[cfg(feature = "file-stream")]
+/// Module for handling file streams.
+pub mod file_stream;
+
+#[cfg(feature = "file-stream")]
+pub use file_stream::FileStream;
+
 #[cfg(feature = "error-response")]
 pub use error_response::InternalServerError;