+++ axum-extra/CHANGELOG.md
@@ -13,12 +13,14 @@ and this project adheres to [Semantic Versioning].
 - **added:** Add `json!` for easy construction of JSON responses ([#2962])
 - **added:** Add `InternalServerError` response for logging an internal error
   and returning HTTP 500 in a convenient way. ([#3010])
+- **added:** Add `FileStream` for easy construction of file stream responses ([#3047])
 
 [#2242]: https://github.com/tokio-rs/axum/pull/2242
 [#3058]: https://github.com/tokio-rs/axum/pull/3058
 [#2961]: https://github.com/tokio-rs/axum/pull/2961
 [#2962]: https://github.com/tokio-rs/axum/pull/2962
 [#3010]: https://github.com/tokio-rs/axum/pull/3010
+[#3047]: https://github.com/tokio-rs/axum/pull/3047
 
 # 0.10.0